+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Prettified golem upgrades particle effects a bit
 * Implemented new scoreboard system, faster and non flicker
 * Fixed NPE when typing /vda setwave without number [#474](https://plajer.xyz/errorservice/viewer.php?id=474)
+* Added Lithuanian and Portuguese Brazilian locales (thanks to POEditor community!)
 
 ### 4.0.2 Release (23.01.2019)
 * Fixed [#415 error](https://plajer.xyz/errorservice/viewer.php?id=415) caused by /vda setprice without price set